[FIX] OCA remarks
