+++ project_model_to_task/README.rst
@@ -2,11 +2,12 @@
 .. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
     :alt: License: AGPL-3
 
-Model to Project Task
+
+Project Model to Task
 =====================
 
-This module allows to end users to create an task from any configured models
-and link automatically the initial object to the task with a reference field.
+This module allows end users to create a task from any configured models
+and automatically link it to the initial object via a reference field.
 
 
 Usage
@@ -14,11 +15,10 @@ Usage
 
 To use this module, you need to:
 
-* go to a model or object in tree or form view configured (ie Partner ou Product).
+* go to a model or object in tree or form view (ie Partner ou Product).
 * select a record (if your are in tree view).
-* select 'Define a task' in the 'More' button.
-* the task and its 'Task Origin' field is set
-* complete and save the task form.
+* select 'Create related task' in the 'More' button.
+* the task and its 'Task Origin' field is set: complete and save the task form.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
@@ -38,7 +38,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/project/issues>`_.
 In case of trouble, please check there if your task has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
-`here <https://github.com/OCA/project/issues/new?body=module:%20model_to_project_task%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`here <https://github.com/OCA/project/issues/new?body=module:%20project_model_to_task%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 
 Credits
@@ -63,4 +63,4 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit http://odoo-community.org.
+To contribute to this module, please visit https://odoo-community.org.
